fix(HomepageLanding): refactor
