+++ src/components/DesktopContainer.js
@@ -1,0 +1,55 @@
+import PropTypes from 'prop-types'
+import React, { Component } from 'react'
+import { connect } from 'react-redux'
+import { startFetch } from '../store'
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+} from 'semantic-ui-react'
+import { Link } from 'react-router-dom';
+import Spinner from './Spinner'
+import HomepageHeading from './HomepageHeading'
+
+
+
+
+class DesktopContainer extends Component {
+    state = {}
+  
+    hideFixedMenu = () => this.setState({ fixed: false })
+    showFixedMenu = () => this.setState({ fixed: true })
+  
+    render() {
+      const { children } = this.props
+      const { fixed } = this.state
+  
+      return (
+        <Responsive {...Responsive.onlyComputer}>
+          <Visibility>
+            <Segment className="landing-header-container" inverted textAlign='center' style={{ minHeight: 700, padding: '1em 0em' }} vertical>
+              <HomepageHeading />
+            </Segment>
+          </Visibility>
+          {children}
+        </Responsive>
+      )
+    }
+  }
+  
+  DesktopContainer.propTypes = {
+    children: PropTypes.node,
+  }
+
+
+  export default DesktopContainer
