+++ src/components/HomepageLayout.js
@@ -19,118 +19,10 @@ import {
 } from 'semantic-ui-react'
 import { Link } from 'react-router-dom';
 import Spinner from './Spinner'
+import ResponsiveContainer from './ResponsiveContainer'
 
-
-// change in-line styling to CSS or styled components before committing
-const HomepageHeading = ({ mobile }) => (
-
-  <Container text>
-
-    <Header
-      as='h1'
-      content='The Network'
-      inverted
-      style={{
-        fontSize: mobile ? '2em' : '4em',
-        fontWeight: 'normal',
-        marginBottom: 0,
-        marginTop: mobile ? '1.5em' : '3em',
-      }}
-    />
-    <Header
-      as='h2'
-      content='We help you stay connected.'
-      inverted
-      style={{
-        fontSize: mobile ? '1.5em' : '1.7em',
-        fontWeight: 'normal',
-        marginTop: mobile ? '0.5em' : '1.5em',
-      }}
-    />
-    <Button className='mouse-cursor-gradient-tracking' primary size='huge' href='/login'>
-      Get Started
-      <Icon name='right arrow' />
-    </Button>
-  </Container>
-)
-
-HomepageHeading.propTypes = {
-  mobile: PropTypes.bool,
-}
-
-class DesktopContainer extends Component {
-  state = {}
-
-  hideFixedMenu = () => this.setState({ fixed: false })
-  showFixedMenu = () => this.setState({ fixed: true })
-
-  render() {
-    const { children } = this.props
-    const { fixed } = this.state
-
-    return (
-      <Responsive {...Responsive.onlyComputer}>
-        <Visibility>
-          <Segment className="landing-header-container" inverted textAlign='center' style={{ minHeight: 700, padding: '1em 0em' }} vertical>
-            <HomepageHeading />
-          </Segment>
-        </Visibility>
-        {children}
-      </Responsive>
-    )
-  }
-}
-
-DesktopContainer.propTypes = {
-  children: PropTypes.node,
-}
-
-class MobileContainer extends Component {
-  state = {}
-
-  handlePusherClick = () => {
-    const { sidebarOpened } = this.state
-
-    if (sidebarOpened) this.setState({ sidebarOpened: false })
-  }
-
-  handleToggle = () => this.setState({ sidebarOpened: !this.state.sidebarOpened })
-
-  render() {
-    const { children } = this.props
-    const { sidebarOpened } = this.state
-
-    return (
-      <Responsive {...Responsive.onlyMobile}>
-
-        <Sidebar.Pusher dimmed={sidebarOpened} onClick={this.handlePusherClick} style={{ minHeight: '100vh' }}>
-          <Segment className="landing-header-container" inverted textAlign='center' style={{ minHeight: 350, padding: '1em 0em' }} vertical>
-            <HomepageHeading mobile />
-          </Segment>
-          {children}
-        </Sidebar.Pusher>
-      </Responsive>
-    )
-  }
-}
-
-MobileContainer.propTypes = {
-  children: PropTypes.node,
-}
-
-const ResponsiveContainer = ({ children }) => (
-  <div>
-    <DesktopContainer>{children}</DesktopContainer>
-    <MobileContainer>{children}</MobileContainer>
-  </div>
-)
-
-ResponsiveContainer.propTypes = {
-  children: PropTypes.node,
-}
 
 class HomepageLayout extends React.Component {
-
 
   render() {
     return (
