+++ src/components/MobileContainer.js
@@ -1,0 +1,59 @@
+import PropTypes from 'prop-types'
+import React, { Component } from 'react'
+import { connect } from 'react-redux'
+import { startFetch } from '../store'
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+} from 'semantic-ui-react'
+import { Link } from 'react-router-dom';
+import Spinner from './Spinner'
+import HomepageHeading from './HomepageHeading'
+
+
+class MobileContainer extends Component {
+    state = {}
+  
+    handlePusherClick = () => {
+      const { sidebarOpened } = this.state
+  
+      if (sidebarOpened) this.setState({ sidebarOpened: false })
+    }
+  
+    handleToggle = () => this.setState({ sidebarOpened: !this.state.sidebarOpened })
+  
+    render() {
+      const { children } = this.props
+      const { sidebarOpened } = this.state
+  
+      return (
+        <Responsive {...Responsive.onlyMobile}>
+  
+          <Sidebar.Pusher dimmed={sidebarOpened} onClick={this.handlePusherClick} style={{ minHeight: '100vh' }}>
+            <Segment className="landing-header-container" inverted textAlign='center' style={{ minHeight: 350, padding: '1em 0em' }} vertical>
+              <HomepageHeading mobile />
+            </Segment>
+            {children}
+          </Sidebar.Pusher>
+        </Responsive>
+      )
+    }
+  }
+  
+  MobileContainer.propTypes = {
+    children: PropTypes.node,
+  }
+
+
+  export default MobileContainer
