+++ src/components/ResponsiveContainer.js
@@ -1,0 +1,36 @@
+import PropTypes from 'prop-types'
+import React, { Component } from 'react'
+import { connect } from 'react-redux'
+import { startFetch } from '../store'
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+} from 'semantic-ui-react'
+import { Link } from 'react-router-dom';
+import Spinner from './Spinner'
+import DesktopContainer from './DestopContainer'
+import MobileContainer from './MobileContainer'
+
+
+const ResponsiveContainer = ({ children }) => (
+    <div>
+      <DesktopContainer>{children}</DesktopContainer>
+      <MobileContainer>{children}</MobileContainer>
+    </div>
+  )
+  
+  ResponsiveContainer.propTypes = {
+    children: PropTypes.node,
+  }
+
